Update Readme File
Fixed documentation of Miwok-Spanish-English-Dictionary

I added information into the Readme so that people know why this dictionary is useful.
I added information regarding that people can learn about numbers, family members, colors and common use phrases. All this can be found in both languages Spanish and English.
In some cases such as numbers and family members, an image is used so a student visualizes that and learn faster.

It was not explained what the dictionary is used for or what features it has.